+++ python/scripts/integration_test_report/__init__.py
@@ -7,5 +7,5 @@ them with historical data, and generates a markdown trend report showing
 per-test status across the last N runs.
 
 Usage:
-    uv run python -m scripts.flaky_report <reports-dir> <history-file> <output-file>
+    uv run python -m scripts.integration_test_report <reports-dir> <history-file> <output-file>
 """
